+++ src/utils/index.ts
@@ -3,4 +3,4 @@ export * from './Amount';
 export * from './Gas';
 export * from './outcome';
 export * from './transform';
-export * from './serde';
+export * from './common';
